+++ src/main/java/com/lipari/app/products/repositories/ProductRepository.java
@@ -5,11 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product,Integer> {
+public interface ProductRepository extends JpaRepository<Product,Long> {
 
-    boolean existsById(Integer id);
+    boolean existsById(Long id);
     boolean existsByCodice(int codice);
 
-    Product getProductById(Integer id);
+    Product getProductById(Long id);
 
 }
