+++ src/main/java/com/lipari/app/users/repositories/AddressRepo.java
@@ -11,17 +11,17 @@ import org.springframework.stereotype.Repository;
 import com.lipari.app.users.entities.Address;
 
 @Repository
-public interface AddressRepo extends JpaRepository<Address, Integer> {
+public interface AddressRepo extends JpaRepository<Address, Long> {
 
 	@Query(value = "SELECT * FROM t_address WHERE user_id = :userId", nativeQuery = true)
-	List<String> getAllUserAddress(@Param("userId") int userId);
+	List<String> getAllUserAddress(@Param("userId") Long userId);
 	
 	
 	@Query(value = "SELECT * FROM t_address WHERE id = :id", nativeQuery = true)
-	Optional<Address> getAddressById(@Param("id") int id);
+	Optional<Address> getAddressById(@Param("id") Long id);
 	
 	@Query(value = "SELECT * FROM t_address WHERE user_id = :user_id AND idnirizzo = :i", nativeQuery = true)
-	Address addressAlreadyExist(@Param("user_id") int id, @Param("i") String indirizzo);
+	Address addressAlreadyExist(@Param("user_id") Long id, @Param("i") String indirizzo);
 	
 
 }
